+++ spark-kernel-dsv2/src/main/java/io/delta/spark/dsv2/scan/batch/KernelSparkBatchScan.java
@@ -17,14 +17,13 @@ package io.delta.spark.dsv2.scan.batch;
 
 import static java.util.Objects.requireNonNull;
 
+import io.delta.spark.dsv2.scan.KernelSparkScanContext;
 import java.io.Serializable;
 import org.apache.spark.sql.connector.read.Batch;
 import org.apache.spark.sql.connector.read.InputPartition;
 import org.apache.spark.sql.connector.read.PartitionReaderFactory;
 
-/**
- * Spark Batch implementation backed by Delta Kernel Scan.
- */
+/** Spark Batch implementation backed by Delta Kernel Scan. */
 public class KernelSparkBatchScan implements Batch, Serializable {
 
   private final KernelSparkScanContext sharedContext;
